+++ psi_cardinality/c/psi_cardinality_server_c.h
@@ -28,7 +28,7 @@ typedef void *psi_cardinality_server_ctx;
 
 struct server_buffer_t {
   const char *buff;
-  size_t buff_len;
+  uint32_t buff_len;
 };
 
 int psi_cardinality_server_create_with_new_key(psi_cardinality_server_ctx *ctx,
@@ -40,17 +40,17 @@ void psi_cardinality_server_delete(psi_cardinality_server_ctx *ctx);
 
 int psi_cardinality_server_create_setup_message(
     psi_cardinality_server_ctx ctx, double fpr, int64_t num_client_inputs,
-    struct server_buffer_t *input, size_t input_len, char **output,
-    size_t *output_len, char **error_out);
+    struct server_buffer_t *input, uint32_t input_len, char **output,
+    uint32_t *output_len, char **error_out);
 void psi_cardinality_server_delete_buffer(psi_cardinality_server_ctx ctx,
                                           char **input);
 
 int psi_cardinality_server_process_request(
     psi_cardinality_server_ctx ctx, struct server_buffer_t client_request,
-    char **output, size_t *output_len, char **error_out);
+    char **output, uint32_t *output_len, char **error_out);
 int psi_cardinality_server_get_private_key_bytes(psi_cardinality_server_ctx ctx,
                                                  char **output,
-                                                 size_t *output_len,
+                                                 uint32_t *output_len,
                                                  char **error_out);
 
 #ifdef __cplusplus
